# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for news.",
        "title": "News API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/news": {
            "get": {
                "description": "Отримати список новин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Список новин",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticlesResponse200"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticlesResponse500"
                        }
                    }
                }
            },
            "post": {
                "description": "Створення новини",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Створення новини",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "description": "news article request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewsArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse400"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse500"
                        }
                    }
                }
            }
        },
        "/news/{id}": {
            "get": {
                "description": "Інформація про новину",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Інформація про новину",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid новини",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse400"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse500"
                        }
                    }
                }
            },
            "put": {
                "description": "Оновлення новини",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "Оновлення новини",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid новини",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "description": "news article request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewsArticleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse500"
                        }
                    }
                }
            },
            "delete": {
                "description": "остаточне видалення новини",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "остаточне видалення новини",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid новини",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse500"
                        }
                    }
                }
            }
        },
        "/news/{id}/recover": {
            "patch": {
                "description": "відновлення новини після м'якого видалення",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "відновлення новини після м'якого видалення",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid новини",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse500"
                        }
                    }
                }
            }
        },
        "/news/{id}/trash": {
            "patch": {
                "description": "м'яке видалення новини",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "news"
                ],
                "summary": "м'яке видалення новини",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid новини",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsArticleResponse500"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.FileUploadDto": {
            "type": "object",
            "properties": {
                "entity_id": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                },
                "entity_type": {
                    "type": "string",
                    "example": "news_articles"
                },
                "name": {
                    "type": "string",
                    "example": "article_img_123.png"
                },
                "path": {
                    "type": "string",
                    "example": "/uploads/article_img_123.png"
                },
                "uuid": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                }
            }
        },
        "models.NewsArticleDTO": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string",
                    "example": "new_article_uri"
                },
                "content": {
                    "type": "string",
                    "example": "Сталось щось добре"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FileUploadDto"
                    }
                },
                "group": {
                    "$ref": "#/definitions/models.NewsGroupDTO"
                },
                "group_id": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                },
                "published": {
                    "type": "boolean",
                    "example": true
                },
                "published_at": {
                    "type": "string",
                    "example": "2024-12-05 12:48"
                },
                "short_description": {
                    "type": "string",
                    "example": "Короткий опис новини."
                },
                "title": {
                    "type": "string",
                    "example": "Хороша новина"
                },
                "user_id": {
                    "type": "integer",
                    "example": 456
                },
                "uuid": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                }
            }
        },
        "models.NewsGroupDTO": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string",
                    "example": "sport"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FileUploadDto"
                    }
                },
                "published": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Спорт"
                },
                "uuid": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                }
            }
        },
...
        "request.NewsArticleRequest": {
            "type": "object",
            "required": [
                "alias",
                "content",
                "title"
            ],
            "properties": {
                "alias": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "nova_stattya"
                },
                "content": {
                    "type": "string",
                    "maxLength": 64000,
                    "example": "Сьогодні щось відбулось."
                },
                "group_id": {
                    "type": "string",
                    "example": "30194cd77-d0ab-74db-88be-f9de341a4b5f"
                },
                "published": {
                    "type": "boolean"
                },
                "published_at": {
                    "type": "string",
                    "example": "2006-01-02T15:04:05Z"
                },
                "short_description": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Короткий опис новини."
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Нова новина"
                }
            }
        },
...
        "response.DocGetNewsArticleResponse200": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/models.NewsArticleDTO"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsArticleResponse400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsArticleResponse404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsArticleResponse500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "щось пішло не так"
                    ]
                }
            }
        },
        "response.DocGetNewsArticlesResponse200": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NewsArticleDTO"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsArticlesResponse500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
    },
...
}